cleaning fitting script and App_Fitting
